# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Prospect API
  version: 0.1.0
  description: Evidence-first financial document intelligence API.
servers:
  - url: /api/v1
security:
  - sessionToken: []

# Every non-2xx response body is an Error. A document owned by another user returns the
# same 404 as a nonexistent one.
paths:
...
  /data-quality:
    get:
      summary: Run data quality rules and list detected anomalies
      parameters:
        - name: document_id
          in: query
          schema:
            type: string
            format: uuid
        - name: company_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved, ignored]
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warning, error]
        - name: rule_code
          in: query
          schema:
            type: string
        - name: run_fresh
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Data quality issues and summary counts
          content:
            application/json:
              schema:
                type: object
                required: [items, summary]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DataQualityIssueOut"
                  summary:
                    $ref: "#/components/schemas/DataQualitySummary"
  /data-quality/{issueId}:
    patch:
      summary: Update data quality anomaly status
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataQualityIssueUpdate"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataQualityIssueOut"
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
...
  schemas:
...
    DataQualityIssueOut:
      type: object
      required: [id, rule_code, severity, status, description, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        document_id:
          type: string
          format: uuid
          nullable: true
        company_id:
          type: string
          format: uuid
          nullable: true
        rule_code:
          type: string
        severity:
          type: string
          enum: [info, warning, error]
        status:
          type: string
          enum: [open, acknowledged, resolved, ignored]
        metric:
          type: string
          nullable: true
        period_label:
          type: string
          nullable: true
        description:
          type: string
        evidence_id:
          type: string
          format: uuid
          nullable: true
        details:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DataQualityIssueUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, acknowledged, resolved, ignored]
    DataQualitySummary:
      type: object
      required: [total, by_severity, by_status]
      properties:
        total:
          type: integer
        by_severity:
          type: object
        by_status:
          type: object
...